jl4-mlir: single-source runtime shared by the CLI and HTTP wrapper
The ~450 lines of ABI marshaling + intrinsic JS were duplicated across
app/Main.hs's generateNodeRunner template and scripts/wasm-server.mjs.
Consolidated into runtime/jl4-runtime.mjs as a single createRuntime()
factory.

- runtime/jl4-runtime.mjs: authoritative runtime — u64<->f64 reinterpret,
  allocator, DATE / TIME / DATETIME helpers, marshaling, unmarshaling,
  Proxy-wrapped env import table, invokeFunction.
- scripts/wasm-server.mjs: drops 400+ lines, just imports createRuntime.
  Output key order matches Aeson (alphabetical) so responses stay
  byte-identical to jl4-service.
- app/Main.hs: embeds runtime/jl4-runtime.mjs at compile time via
  file-embed. generateNodeRunner shrinks to a launcher that reads the
  wasm, calls createRuntime, invokes the function. Node now runs with
  --input-type=module so the embedded 'export' is legal.
- jl4-mlir.cabal: adds file-embed dep; declares the runtime as
  extra-source-files so TH re-embeds after edits.

Verified against jl4-service on the datetime and intrinsics probes via
both the CLI and the HTTP wrapper — byte-identical SimpleResponse
payloads. Batch-compile regression: still 0 failures.
