+++ jl4-mlir/app/Main.hs
@@ -1,4 +1,5 @@
 {-# LANGUAGE OverloadedStrings #-}
+{-# LANGUAGE TemplateHaskell #-}
 
 -- | @jl4-mlir@ CLI — compile and run L4 programs via MLIR/WASM.
 --
@@ -25,9 +26,11 @@ module Main (main) where
 import Control.Monad (when, unless)
 import qualified Data.Aeson as Aeson
 import qualified Data.Aeson.Encode.Pretty as Aeson.Pretty
+import qualified Data.Aeson.Key as Aeson.Key
 import qualified Data.Aeson.KeyMap as Aeson.KeyMap
 import qualified Data.ByteString.Lazy as LBS
 import qualified Data.ByteString.Lazy.Char8 as LBSC
+import Data.FileEmbed (embedStringFile)
 import Data.List (isSuffixOf)
 import Data.Text (Text)
 import qualified Data.Text as Text
@@ -311,6 +314,13 @@ findExport name bundle =
 -- Runtime: Node.js
 -- ---------------------------------------------------------------------------
 
+-- | Contents of @runtime/jl4-runtime.mjs@, baked into the binary at
+-- compile time. The same file is imported by @scripts/wasm-server.mjs@
+-- so the CLI and the HTTP wrapper share a single source of truth for
+-- the ABI marshaling and the @env.__l4_*@ import table.
+runtimeJs :: String
+runtimeJs = $(embedStringFile "runtime/jl4-runtime.mjs")
+
 runNode :: FilePath -> FunctionExport -> ResolvedArgs -> IO (Either Text ResponseWithReason)
 runNode wasmPath fe resolved = do
   mTool <- findExecutable "node"
@@ -318,7 +328,8 @@ runNode wasmPath fe resolved = do
     Nothing -> pure $ Left "node not found in PATH"
     Just node -> do
       let jsCode = generateNodeRunner wasmPath fe resolved
-      (exitCode, stdout_, stderr_) <- readProcessWithExitCode node ["-e", jsCode] ""
+      (exitCode, stdout_, stderr_) <-
+        readProcessWithExitCode node ["--input-type=module", "-e", jsCode] ""
       case exitCode of
         ExitSuccess ->
           case Aeson.eitherDecode (LBSC.pack stdout_) of
@@ -328,475 +339,39 @@ runNode wasmPath fe resolved = do
             Right r -> pure $ Right r
         ExitFailure _ -> pure $ Left $ Text.pack stderr_
 
--- | Generate the Node.js runner script.
+
+-- | Generate the per-invocation Node.js script.
 --
--- The script loads the .wasm, supplies stubs for the @env.__l4_*@
--- runtime imports, and marshals each positional argument into a WASM
--- value based on its schema type. Strings become pointers to UTF-8
--- bytes in linear memory, records become laid-out structs, and lists
--- become linked-list nodes.
---
--- The final result is printed as a JSON @ResponseWithReason@ document.
+-- Emits an ES module containing the embedded runtime (from
+-- @runtime/jl4-runtime.mjs@) followed by a short launcher that loads
+-- the .wasm, instantiates it with the runtime's import table, and
+-- invokes one exported function. Shared with
+-- @scripts/wasm-server.mjs@ via the single runtime source file.
 generateNodeRunner :: FilePath -> FunctionExport -> ResolvedArgs -> String
 generateNodeRunner wasmPath fe resolved =
-  let argValuesJson =
-        LBSC.unpack $ Aeson.encode $ Aeson.toJSON
-          [ argJson v | (_, _, v) <- resolved.raPositional ]
-      schemaTypesJson =
-        LBSC.unpack $ Aeson.encode $ Aeson.toJSON
-          [ p | (_, p, _) <- resolved.raPositional ]
+  let metaJson = LBSC.unpack (Aeson.encode (Aeson.toJSON fe))
+      -- Build @{paramName: value}@ for the positional args. Params
+      -- the caller omitted become @null@ so the runtime can emit a
+      -- NOTHING MAYBE or a zero for simple scalars as appropriate.
+      argsMap =
+        Aeson.Object $ Aeson.KeyMap.fromList
+          [ (Aeson.Key.fromText pname, argJson v)
+          | (pname, _, v) <- resolved.raPositional
+          ]
+      argsJson = LBSC.unpack (Aeson.encode argsMap)
   in unlines
-    [ "const fs = require('fs');"
+    [ runtimeJs
+    , "import fs from 'node:fs';"
     , "const buf = fs.readFileSync(" <> show wasmPath <> ");"
-    , ""
-    , "// Scratch buffer for u64<->f64 bit-pattern conversion. The compiler's"
-    , "// uniform f64 ABI encodes pointers as f64 values whose raw 8 bytes are"
-    , "// the u64 pointer — *not* numerically equal. JS->wasm must pass those"
-    , "// bits through an f64 register."
-    , "const scratch = new ArrayBuffer(8);"
-    , "const sView = new DataView(scratch);"
-    , "function u64ToF64(u) { sView.setBigUint64(0, BigInt(u), true); return sView.getFloat64(0, true); }"
-    , "function f64ToU64(f) { sView.setFloat64(0, Number(f), true); return sView.getBigUint64(0, true); }"
-    , ""
-    , "let memory, memView, memU8;"
-    , "let heapPtr = 1024;"
-    , "function allocBytes(n) {"
-    , "  const nn = Number(n);"
-    , "  const p = heapPtr;"
-    , "  heapPtr += (nn + 7) & ~7;"
-    , "  return p;"
-    , "}"
-    , ""
-    , "function readCString(p) {"
-    , "  const ptr = Number(p);"
-    , "  if (!ptr) return '';"
-    , "  let end = ptr;"
-    , "  while (end < memU8.length && memU8[end] !== 0) end++;"
-    , "  return Buffer.from(memU8.subarray(ptr, end)).toString('utf8');"
-    , "}"
-    , ""
-    , "function writeString(s) {"
-    , "  const bytes = Buffer.from(String(s), 'utf8');"
-    , "  const p = allocBytes(bytes.length + 1);"
-    , "  memU8.set(bytes, p);"
-    , "  memU8[p + bytes.length] = 0;"
-    , "  return p;"
-    , "}"
-    , ""
-    , "// --- Date / Time / Datetime helpers ---"
-    , "// DATE: f64 days with a Sunday-zero offset. The raw number is"
-    , "//   days-since-1969-12-28 (the last Sunday before the Unix"
-    , "//   epoch), so 'date % 7' gives the day-of-week with Sunday=0,"
-    , "//   matching the library's weekday constants."
-    , "// TIME: f64 seconds since midnight"
-    , "// DATETIME: pointer (bit-cast to f64) to a 2-slot record [utcSeconds, tzNamePtr]"
-    , "const DATE_EPOCH_SHIFT = 4;  // days from 1969-12-28 (Sun) to 1970-01-01 (Thu)"
-    , "function dateFromDMY(d, m, y) { return Math.floor(Date.UTC(y, m - 1, d) / 86400000) + DATE_EPOCH_SHIFT; }"
-    , "function dateToUnixDays(x) { return Number(x) - DATE_EPOCH_SHIFT; }"
-    , "function parseDateStr(s) {"
-    , "  const m = /^(\\d{4})-(\\d{2})-(\\d{2})/.exec(String(s));"
-    , "  if (!m) return NaN;"
-    , "  return dateFromDMY(+m[3], +m[2], +m[1]);"
-    , "}"
-    , "function formatDate(days) {"
-    , "  const d = new Date(dateToUnixDays(days) * 86400000);"
-    , "  const Y = d.getUTCFullYear(), M = d.getUTCMonth() + 1, D = d.getUTCDate();"
-    , "  return String(Y).padStart(4, '0') + '-' + String(M).padStart(2,'0') + '-' + String(D).padStart(2,'0');"
-    , "}"
-    , "function timeFromHMS(h, m, s) { return h * 3600 + m * 60 + s; }"
-    , "function parseTimeStr(s) {"
-    , "  const m = /^(\\d{1,2}):(\\d{2})(?::(\\d{2}(?:\\.\\d+)?))?/.exec(String(s));"
-    , "  if (!m) return NaN;"
-    , "  return (+m[1]) * 3600 + (+m[2]) * 60 + (+(m[3] || 0));"
-    , "}"
-    , "function formatTime(sec) {"
-    , "  const t = Number(sec);"
-    , "  const h = Math.floor(t / 3600), mi = Math.floor((t % 3600) / 60), s = Math.floor(t % 60);"
-    , "  return String(h).padStart(2,'0') + ':' + String(mi).padStart(2,'0') + ':' + String(s).padStart(2,'0');"
-    , "}"
-    , "function tzOffsetMs(tz, utcMs) {"
-    , "  if (!tz || tz === 'UTC' || tz === 'Etc/UTC') return 0;"
-    , "  const dtf = new Intl.DateTimeFormat('en-US', {"
-    , "    timeZone: tz, hourCycle: 'h23',"
-    , "    year: 'numeric', month: 'numeric', day: 'numeric',"
-    , "    hour: 'numeric', minute: 'numeric', second: 'numeric',"
-    , "  });"
-    , "  const parts = dtf.formatToParts(new Date(utcMs));"
-    , "  const map = {};"
-    , "  for (const p of parts) if (p.type !== 'literal') map[p.type] = parseInt(p.value, 10);"
-    , "  const asUTC = Date.UTC(map.year, map.month - 1, map.day, map.hour, map.minute, map.second);"
-    , "  return asUTC - utcMs;"
-    , "}"
-    , "function datetimeFromDTZ(daysF, secsF, tzPtrF) {"
-    , "  const tzStr = readCString(Number(f64ToU64(tzPtrF))) || 'UTC';"
-    , "  const localMs = dateToUnixDays(daysF) * 86400000 + Number(secsF) * 1000;"
-    , "  const utcMs = localMs - tzOffsetMs(tzStr, localMs);"
-    , "  const p = allocBytes(16);"
-    , "  memView.setFloat64(p, utcMs / 1000, true);"
-    , "  memView.setBigUint64(p + 8, BigInt(writeString(tzStr)), true);"
-    , "  return u64ToF64(p);"
-    , "}"
-    , "function formatDatetime(dtPtrF) {"
-    , "  const p = Number(f64ToU64(dtPtrF));"
-    , "  if (!p) return '';"
-    , "  const utcSec = memView.getFloat64(p, true);"
-    , "  const d = new Date(utcSec * 1000);"
-    , "  return String(d.getUTCFullYear()).padStart(4,'0') + '-' + String(d.getUTCMonth()+1).padStart(2,'0') + '-' + String(d.getUTCDate()).padStart(2,'0')"
-    , "       + ' ' + String(d.getUTCHours()).padStart(2,'0') + ':' + String(d.getUTCMinutes()).padStart(2,'0') + ':' + String(d.getUTCSeconds()).padStart(2,'0')"
-    , "       + ' UTC';"
-    , "}"
-    , "// JUST-wrapped result helper used by __l4_to_date / __l4_to_time / __l4_to_datetime."
-    , "function mbJustF64(f64Val) {"
-    , "  const p = allocBytes(16);"
-    , "  memView.setFloat64(p, 1.0, true);"
-    , "  memView.setFloat64(p + 8, Number(f64Val), true);"
-    , "  return u64ToF64(p);"
-    , "}"
-    , "function mbNothing() {"
-    , "  const p = allocBytes(16);"
-    , "  memView.setFloat64(p, 0.0, true);"
-    , "  memView.setFloat64(p + 8, 0.0, true);"
-    , "  return u64ToF64(p);"
-    , "}"
-    , ""
-    , "// Is this field/value a pointer under the uniform f64 ABI?"
-    , "// Numbers and booleans are numeric f64; strings/objects/arrays/enums are"
-    , "// pointers (stored in memory as raw u64 bit-patterns, crossed across the"
-    , "// wasm ABI boundary via the scratch reinterpret cast)."
-    , "function isPointerSchema(schema) {"
-    , "  if (!schema) return false;"
-    , "  if (schema.enum && schema.enum.length > 0) return false;  // enum tags are numeric i32"
-    , "  // DATE (format=date) and TIME (format=time) wire as strings but"
-    , "  // the compiled code expects a numeric f64 (days / seconds)."
-    , "  if (schema.format === 'date' || schema.format === 'time') return false;"
-    , "  const t = schema.type;"
-    , "  return t === 'string' || t === 'array' || t === 'object';"
-    , "}"
-    , ""
-    , "function marshalArg(value, schema) {"
-    , "  if (value === null || value === undefined) return 0;"
-    , "  if (schema && schema.enum && schema.enum.length > 0) {"
-    , "    // Enum: index into the enum list (matches compiler's i32 tag)"
-    , "    const idx = schema.enum.indexOf(String(value));"
-    , "    return idx >= 0 ? idx : 0;"
-    , "  }"
-    , "  // DATE / TIME / DATETIME: wire format is a string; internal is numeric"
-    , "  if (schema && schema.format === 'date' && schema.type === 'string') {"
-    , "    return parseDateStr(value);"
-    , "  }"
-    , "  if (schema && schema.format === 'time' && schema.type === 'string') {"
-    , "    return parseTimeStr(value);"
-    , "  }"
-    , "  switch (schema && schema.type) {"
-    , "    case 'number':  return Number(value);"
-    , "    case 'boolean': return value ? 1 : 0;"
-    , "    case 'string':  return writeString(value);"
-    , "    case 'array':   return marshalList(value, (schema && schema.items) || {});"
-    , "    case 'object':  return marshalStruct(value, schema);"
-    , "    default: return Number(value) || 0;"
-    , "  }"
-    , "}"
-    , ""
-    , "// Allocate a 2-slot MAYBE record matching the compiler's convention:"
-    , "//   slot 0 = tag (1.0 = JUST, 0.0 = NOTHING)"
-    , "//   slot 1 = payload, encoded per inner type"
-    , "function marshalMaybe(value, innerSchema) {"
-    , "  const p = allocBytes(16);"
-    , "  if (value === null || value === undefined) {"
-    , "    memView.setFloat64(p, 0.0, true);"
-    , "    memView.setFloat64(p + 8, 0.0, true);"
-    , "    return p;"
-    , "  }"
-    , "  memView.setFloat64(p, 1.0, true);"
-    , "  const inner = marshalArg(value, innerSchema);"
-    , "  if (innerSchema && innerSchema.type === 'number') {"
-    , "    memView.setFloat64(p + 8, Number(inner), true);"
-    , "  } else {"
-    , "    memView.setBigUint64(p + 8, BigInt(inner), true);"
-    , "  }"
-    , "  return p;"
-    , "}"
-    , ""
-    , "function marshalStruct(value, schema) {"
-    , "  if (!schema || !schema.properties) return 0;"
-    , "  const order = schema.propertyOrder || Object.keys(schema.properties);"
-    , "  const required = new Set(schema.required || order);"
-    , "  const size = order.length * 8;"
-    , "  const p = allocBytes(size);"
-    , "  order.forEach((name, idx) => {"
-    , "    const fieldSchema = schema.properties[name] || {};"
-    , "    const v = (value || {})[name];"
-    , "    const isMaybe = !required.has(name);"
-    , "    const isEnum = !!(fieldSchema.enum && fieldSchema.enum.length > 0);"
-    , "    if (isMaybe) {"
-    , "      // Optional field in the record — L4 MAYBE. Always emit a 2-slot"
-    , "      // struct and store its pointer in this field."
-    , "      const mbPtr = marshalMaybe(v, fieldSchema);"
-    , "      memView.setBigUint64(p + idx * 8, BigInt(mbPtr), true);"
-    , "    } else if (fieldSchema.type === 'number' || isEnum) {"
-    , "      // Number or enum tag — numeric f64 in the uniform ABI."
-    , "      const m = marshalArg(v, fieldSchema);"
-    , "      memView.setFloat64(p + idx * 8, Number(m), true);"
-    , "    } else {"
-    , "      // Pointer-typed (string/object/array) or boolean — store the raw"
-    , "      // u64 bit pattern. wasm reads as f64, bitcasts back to recover"
-    , "      // the pointer (or compares against 0.0 for bool)."
-    , "      const m = marshalArg(v, fieldSchema);"
-    , "      memView.setBigUint64(p + idx * 8, BigInt(m), true);"
-    , "    }"
-    , "  });"
-    , "  return p;"
-    , "}"
-    , ""
-    , "function marshalList(values, itemSchema) {"
-    , "  if (!Array.isArray(values) || values.length === 0) return 0;"
-    , "  let head = 0;"
-    , "  for (let i = values.length - 1; i >= 0; i--) {"
-    , "    const p = allocBytes(16);"
-    , "    const item = marshalArg(values[i], itemSchema);"
-    , "    if (itemSchema && itemSchema.type === 'number') {"
-    , "      memView.setFloat64(p, Number(item), true);"
-    , "    } else {"
-    , "      memView.setBigUint64(p, BigInt(item), true);"
-    , "    }"
-    , "    memView.setBigUint64(p + 8, BigInt(head), true);"
-    , "    head = p;"
-    , "  }"
-    , "  return head;"
-    , "}"
-    , ""
-    , "const imports = {"
-    , "  env: {"
-    , "    __l4_pow:   Math.pow,"
-    , "    __l4_min:   Math.min,"
-    , "    __l4_max:   Math.max,"
-    , "    __l4_abs:   Math.abs,"
-    , "    __l4_floor: Math.floor,"
-    , "    __l4_ceil:  Math.ceil,"
-    , "    __l4_round: Math.round,"
-    , "    __l4_str_concat: (a, b) => a,"
-    , "    __l4_str_eq:     (a, b) => (a === b) ? 1 : 0,"
-    , "    __l4_str_len:    (_p) => 0,"
-    , "    __l4_to_string:  (_n) => 0,"
-    , "    __l4_list_count: (ptr) => {"
-    , "      let n = 0, p = Number(f64ToU64(ptr));"
-    , "      while (p !== 0 && n < 100000) {"
-    , "        p = Number(memView.getBigUint64(p + 8, true));"
-    , "        n++;"
-    , "      }"
-    , "      return n;"
-    , "    },"
-    , "    __l4_list_empty: () => 0,"
-    , "    __l4_alloc: (n) => allocBytes(n),"
-    , "    __l4_free:  (_p) => {},"
-    , "    // libm functions llc emits as calls when lowering MLIR arith"
-    , "    // operations that don't have a native wasm32 instruction."
-    , "    fmod:  (a, b) => Number(a) % Number(b),"
-    , "    fmodf: (a, b) => Number(a) % Number(b),"
-    , "    remainder:  (a, b) => Number(a) - Math.round(Number(a) / Number(b)) * Number(b),"
-    , "    remainderf: (a, b) => Number(a) - Math.round(Number(a) / Number(b)) * Number(b),"
-    , "    trunc: Math.trunc,"
-    , "    // DATE primitives (f64 days since Unix epoch)"
-    , "    __l4_date_from_dmy:    (d, m, y) => dateFromDMY(Number(d), Number(m), Number(y)),"
-    , "    // Our internal DATE rep is already Sunday-aligned, so"
-    , "    // DATE_SERIAL / DATE_FROM_SERIAL are identity."
-    , "    __l4_date_from_serial: (n) => Number(n),"
-    , "    __l4_date_serial:      (d) => Number(d),"
-    , "    __l4_date_day:         (d) => new Date(dateToUnixDays(d) * 86400000).getUTCDate(),"
-    , "    __l4_date_month:       (d) => new Date(dateToUnixDays(d) * 86400000).getUTCMonth() + 1,"
-    , "    __l4_date_year:        (d) => new Date(dateToUnixDays(d) * 86400000).getUTCFullYear(),"
-    , "    __l4_datevalue:        (sPtrF) => {"
-    , "      const s = readCString(Number(f64ToU64(sPtrF)));"
-    , "      const v = parseDateStr(s);"
-    , "      if (!Number.isFinite(v)) { return mbNothing(); }"
-    , "      return mbJustF64(v);"
-    , "    },"
-    , "    __l4_to_date:          (sPtrF) => {"
-    , "      const s = readCString(Number(f64ToU64(sPtrF)));"
-    , "      const v = parseDateStr(s);"
-    , "      if (!Number.isFinite(v)) { return mbNothing(); }"
-    , "      return mbJustF64(v);"
-    , "    },"
-    , "    // TIME primitives (f64 seconds since midnight)"
-    , "    __l4_time_from_hms:    (h, m, s) => timeFromHMS(Number(h), Number(m), Number(s)),"
-    , "    __l4_time_from_serial: (n) => Number(n),"
-    , "    __l4_time_serial:      (t) => Number(t),"
-    , "    __l4_time_hour:        (t) => Math.floor(Number(t) / 3600),"
-    , "    __l4_time_minute:      (t) => Math.floor((Number(t) % 3600) / 60),"
-    , "    __l4_time_second:      (t) => Number(t) % 60,"
-    , "    __l4_to_time:          (sPtrF) => {"
-    , "      const s = readCString(Number(f64ToU64(sPtrF)));"
-    , "      const v = parseTimeStr(s);"
-    , "      if (!Number.isFinite(v)) { return mbNothing(); }"
-    , "      return mbJustF64(v);"
-    , "    },"
-    , "    // DATETIME primitives (pointer to [utcSeconds, tzName])"
-    , "    __l4_datetime_from_dtz: (d, t, tz) => datetimeFromDTZ(d, t, tz),"
-    , "    __l4_datetime_date:     (dtF) => {"
-    , "      const p = Number(f64ToU64(dtF));"
-    , "      if (!p) return 0;"
-    , "      return Math.floor(memView.getFloat64(p, true) / 86400) + DATE_EPOCH_SHIFT;"
-    , "    },"
-    , "    __l4_datetime_time:     (dtF) => {"
-    , "      const p = Number(f64ToU64(dtF));"
-    , "      if (!p) return 0;"
-    , "      const sec = memView.getFloat64(p, true);"
-    , "      return ((sec % 86400) + 86400) % 86400;"
-    , "    },"
-    , "    __l4_datetime_tz:       (dtF) => {"
-    , "      const p = Number(f64ToU64(dtF));"
-    , "      if (!p) return u64ToF64(writeString(''));"
-    , "      const tzPtr = Number(memView.getBigUint64(p + 8, true));"
-    , "      return u64ToF64(tzPtr);"
-    , "    },"
-    , "    __l4_datetime_serial:   (dtF) => {"
-    , "      const p = Number(f64ToU64(dtF));"
-    , "      if (!p) return 0;"
-    , "      return memView.getFloat64(p, true);"
-    , "    },"
-    , "    __l4_to_datetime:       (sPtrF) => {"
-    , "      const s = readCString(Number(f64ToU64(sPtrF)));"
-    , "      const t = Date.parse(s);"
-    , "      if (Number.isNaN(t)) return mbNothing();"
-    , "      const p = allocBytes(16);"
-    , "      memView.setFloat64(p, t / 1000, true);"
-    , "      memView.setBigUint64(p + 8, BigInt(writeString('UTC')), true);"
-    , "      const recPtr = u64ToF64(p);"
-    , "      const jp = allocBytes(16);"
-    , "      memView.setFloat64(jp, 1.0, true);"
-    , "      memView.setFloat64(jp + 8, recPtr, true);"
-    , "      return u64ToF64(jp);"
-    , "    },"
-    , "    // Extended numeric intrinsics."
-    , "    __l4_sqrt:  Math.sqrt,"
-    , "    __l4_ln:    Math.log,"
-    , "    __l4_log10: Math.log10,"
-    , "    __l4_sin:   Math.sin,"
-    , "    __l4_cos:   Math.cos,"
-    , "    __l4_tan:   Math.tan,"
-    , "    __l4_asin:  Math.asin,"
-    , "    __l4_acos:  Math.acos,"
-    , "    __l4_atan:  Math.atan,"
-    , "    __l4_trunc: (x, d) => { const m = Math.pow(10, Number(d)); return Math.trunc(Number(x) * m) / m; },"
-    , "    __l4_is_integer: (x) => Number.isInteger(Number(x)) ? 1 : 0,"
-    , "    // String intrinsics. All string args arrive as f64-bitcast pointers."
-    , "    __l4_string_length: (sF) => readCString(Number(f64ToU64(sF))).length,"
-    , "    __l4_to_upper:      (sF) => u64ToF64(writeString(readCString(Number(f64ToU64(sF))).toUpperCase())),"
-    , "    __l4_to_lower:      (sF) => u64ToF64(writeString(readCString(Number(f64ToU64(sF))).toLowerCase())),"
-    , "    __l4_trim:          (sF) => u64ToF64(writeString(readCString(Number(f64ToU64(sF))).trim())),"
-    , "    __l4_contains:      (hF, nF) => readCString(Number(f64ToU64(hF))).includes(readCString(Number(f64ToU64(nF)))) ? 1 : 0,"
-    , "    __l4_starts_with:   (hF, nF) => readCString(Number(f64ToU64(hF))).startsWith(readCString(Number(f64ToU64(nF)))) ? 1 : 0,"
-    , "    __l4_ends_with:     (hF, nF) => readCString(Number(f64ToU64(hF))).endsWith(readCString(Number(f64ToU64(nF)))) ? 1 : 0,"
-    , "    __l4_index_of:      (hF, nF) => readCString(Number(f64ToU64(hF))).indexOf(readCString(Number(f64ToU64(nF)))),"
-    , "    __l4_char_at:       (sF, iF) => u64ToF64(writeString(readCString(Number(f64ToU64(sF))).charAt(Number(iF)))),"
-    , "    __l4_substring:     (sF, iF, jF) => u64ToF64(writeString(readCString(Number(f64ToU64(sF))).substring(Number(iF), Number(jF)))),"
-    , "    __l4_replace:       (sF, aF, bF) => {"
-    , "      const s = readCString(Number(f64ToU64(sF)));"
-    , "      const a = readCString(Number(f64ToU64(aF)));"
-    , "      const b = readCString(Number(f64ToU64(bF)));"
-    , "      return u64ToF64(writeString(s.split(a).join(b)));"
-    , "    },"
-    , "    __l4_to_number:     (sF) => {"
-    , "      const n = Number(readCString(Number(f64ToU64(sF))));"
-    , "      return Number.isNaN(n) ? mbNothing() : mbJustF64(n);"
-    , "    },"
-    , "    // Nullary temporal. TODAY and NOW return Unix serials; CURRENTTIME"
-    , "    // returns a TIME (seconds-since-midnight UTC)."
-    , "    __l4_today_serial: () => Math.floor(Date.now() / 86400000) + DATE_EPOCH_SHIFT,"
-    , "    __l4_now_serial:   () => Date.now() / 1000,"
-    , "    __l4_current_time: () => { const d = new Date(); return d.getUTCHours()*3600 + d.getUTCMinutes()*60 + d.getUTCSeconds(); },"
-    , "    __l4_timezone_name: () => u64ToF64(writeString('UTC')),"
-    , "    // JSON."
-    , "    __l4_json_encode: (vF) => u64ToF64(writeString(String(Number(vF)))),"
-    , "    __l4_json_decode: (sF) => {"
-    , "      try { const v = JSON.parse(readCString(Number(f64ToU64(sF)))); return typeof v === 'number' ? mbJustF64(v) : mbNothing(); }"
-    , "      catch { return mbNothing(); }"
-    , "    },"
-    , "  }"
-    , "};"
-    , ""
-    , "// Unsupported higher-order / unresolved library names appear as"
-    , "// MLIR externs; Node's WebAssembly.instantiate requires every"
-    , "// declared import to be present. Wrap env in a Proxy that returns"
-    , "// a throwing stub for anything we didn't implement — modules that"
-    , "// never actually call those symbols instantiate fine."
-    , "const envProxy = new Proxy(imports.env, {"
-    , "  get(target, prop) {"
-    , "    if (prop in target) return target[prop];"
-    , "    return () => { throw new Error('l4: unresolved runtime symbol \"' + String(prop) + '\" (higher-order / unsupported library feature)'); };"
-    , "  }"
-    , "});"
-    , ""
-    , "WebAssembly.instantiate(buf, { env: envProxy }).then(({instance}) => {"
-    , "  memory  = instance.exports.memory;"
-    , "  memView = new DataView(memory.buffer);"
-    , "  memU8   = new Uint8Array(memory.buffer);"
-    , ""
-    , "  const fn = instance.exports[" <> show (Text.unpack fe.wasmSymbol) <> "];"
-    , "  if (!fn) {"
-    , "    console.error('WASM export not found: " <> Text.unpack fe.wasmSymbol <> "');"
-    , "    console.error('Available exports:', Object.keys(instance.exports).join(', '));"
-    , "    process.exit(1);"
-    , "  }"
-    , ""
-    , "  const argValues = " <> argValuesJson <> ";"
-    , "  const argSchemas = " <> schemaTypesJson <> ";"
-    , ""
-    , "  // For pointer-typed args: JS number is a u64 pointer; must cross the"
-    , "  // wasm f64 ABI boundary as the matching bit pattern, not numerically."
-    , "  const marshaled = argValues.map((v, i) => {"
-    , "    const m = marshalArg(v, argSchemas[i]);"
-    , "    return isPointerSchema(argSchemas[i]) ? u64ToF64(m) : Number(m);"
-    , "  });"
-    , "  const raw = fn(...marshaled);"
-    , "  const result = unmarshalResult(raw, " <> show (Text.unpack fe.returnType) <> ");"
-    , "  // Emit the inner ResponseWithReason shape. The Haskell driver"
-    , "  // wraps it in {tag:'SimpleResponse', contents:...} before printing."
-    , "  console.log(JSON.stringify({ result: { value: result } }));"
-    , "}).catch(err => {"
-    , "  console.error('WASM error:', err.message || err);"
-    , "  process.exit(1);"
-    , "});"
-    , ""
-    , "function unmarshalScalar(raw, type) {"
-    , "  if (type === 'BOOLEAN')  return Number(raw) !== 0;"
-    , "  if (type === 'NUMBER')   return Number(raw);"
-    , "  if (type === 'STRING')   return readCString(Number(f64ToU64(raw)));"
-    , "  if (type === 'DATE')     return formatDate(raw);"
-    , "  if (type === 'TIME')     return formatTime(raw);"
-    , "  if (type === 'DATETIME') return formatDatetime(raw);"
-    , "  return Number(raw);"
-    , "}"
-    , ""
-    , "// MAYBE values round-trip as pointer-f64 to a 2-slot struct:"
-    , "//   slot 0 = tag (0.0 = NOTHING, 1.0 = JUST)"
-    , "//   slot 1 = payload (f64, interpreted per inner type)"
-    , "function unmarshalMaybe(raw, innerType) {"
-    , "  const ptr = Number(f64ToU64(raw));"
-    , "  if (!ptr) return 'NOTHING';"
-    , "  const tag = memView.getFloat64(ptr, true);"
-    , "  if (tag === 0.0) return 'NOTHING';"
-    , "  // slot 1 holds the payload as an f64"
-    , "  let payload;"
-    , "  if (innerType === 'NUMBER') {"
-    , "    payload = memView.getFloat64(ptr + 8, true);"
-    , "  } else if (innerType === 'BOOLEAN') {"
-    , "    payload = memView.getFloat64(ptr + 8, true) !== 0;"
-    , "  } else if (innerType === 'STRING') {"
-    , "    const sp = Number(memView.getBigUint64(ptr + 8, true));"
-    , "    payload = readCString(sp);"
-    , "  } else {"
-    , "    payload = memView.getFloat64(ptr + 8, true);"
-    , "  }"
-    , "  return { JUST: [payload] };"
-    , "}"
-    , ""
-    , "function unmarshalResult(raw, returnType) {"
-    , "  if (returnType && returnType.indexOf('MAYBE ') === 0) {"
-    , "    return unmarshalMaybe(raw, returnType.slice(6));"
-    , "  }"
-    , "  return unmarshalScalar(raw, returnType);"
-    , "}"
+    , "const meta = " <> metaJson <> ";"
+    , "const args = " <> argsJson <> ";"
+    , "const rt = createRuntime();"
+    , "const { instance } = await WebAssembly.instantiate(buf, rt.makeImports());"
+    , "rt.attachMemory(instance.exports.memory);"
+    , "const result = rt.invokeFunction(instance, meta, args);"
+    , "// Emit the inner ResponseWithReason shape; the Haskell driver"
+    , "// wraps it in the SimpleResponse envelope before printing."
+    , "console.log(JSON.stringify({ result: { value: result } }));"
     ]
 
 -- | Convert a positional argument (with optional value) to a JSON value.
